move cache preparation onto strategy and avoid import loop
