+++ taal/kaiso/manager.py
@@ -5,12 +5,14 @@ from kaiso.exceptions import DeserialisationError
 from kaiso.persistence import Manager as KaisoManager
 from kaiso.types import PersistableType, get_type_id
 
-from taal import (
-    TranslatableString as TaalTranslatableString, is_translatable_value)
 from taal.constants import PLACEHOLDER
 from taal.exceptions import NoTranslatorRegistered
 from taal.kaiso import TranslatableString, TYPE_CONTEXT
 from taal.kaiso.types import get_context, get_message_id
+from taal.translatablestring import (
+    is_translatable_value,
+    TranslatableString as TaalTranslatableString,
+)
 
 
 MISSING = object()
